revert api changes from 6e9aaf0fce0214591225d1fa8aeafb8f5aa802cb
* append linke_turb_forms.py to atmosphere.py
* rm -rf atmosphere/
